Enhanced formatting control for fibers and payloads
Added __format__ methods for fibers and payloads. The
format spec for a fiber is a follows:

[(<coord spec>,<payload spec>)][n][*]

where:
    <coord spec> is an ordinary spec for the coordinates
    <payload spec> is an ordinary spec for the payload
    "n" means add newlines for each subfiber
    "*" means do not truncate with elipsis

The current default behavior is to print on a single line with
an elipsis (...) after two elements in each fiber.

The format spec for a payload is an ordinary spec to use for the
value of the payload.

Currently there is no format control for tensors or ranks, and
they currently print each fiber with newlines and without truncation.