Added methods to transform a Tensor
Added the following methods:

- splitUniform
- splitNonUniform
- splitEqual
- splitUnEqual
- swapRanks
- flattenRanks
- unflattenRanks

All do a "deep copy" of the current Tensor and create a new tensor
maniputed according to the specified transform.

Also added an IPython example and test suite.

Finally, changed Fiber.updateCoords so that the lambda function
is given the current position, coordinate and payload of the element
for which a new coordinate is to be generated.