Refactored image display classes
Repurposed the TensorImage class to be a wrapper for the various
tensor image style classes: TreeImage and UncompressedImage. The
TensorImage class takes a 'style=' keyword paraemter that selects
among the following styles:

- tree - the orginal fiber-tree style (TreeImage)
- uncompressed - an uncompressed style (UncompressedImage)
- tree+uncompressed - a combined image of both the above styles (NEW)

Enhanced TensorCanvas class to include a 'style=' keyword that
picks a style for the images in the movie.

Updated prelude.py to support use of the above styles, and also added
a "run all cells below" button.

TBD: Enhance appearance of the UncompressedImage style.