Label ranks in uncompressed image display
There are some still some anomolies in the display of tensors in an
uncompressed representation. They seem to be related to cases where
final coordinates or entire fibers happen to be empty. The solution is
probably to have the objects carry around some more information like
maxCoord as a rank-level constant rather than trying to compute it
from a the fiber one is currently working on.

Some minor fixes in Eyeriss V2...