Avoid use of Tensor(<YAMLfile>) constructor
Begin deprecating use of the Tensor constructor of the form
Tensor(<YAMLfile>) in favor of Tensor.fromYAMLfile(<YAMLfile>)
by changing all the uses of Tensor(<YAMLfile>) I could find.

This is so that all initialized tensor constructors are of the form
Tensor.from<something>(..) and the raw Tensor constructor can be
eventually be used to create an empty tensor in the cleaner form
Tensor(<rank_ids>, ...).

Still to be done is to treat the first argument of the Tensor()
constructor as a rank_id and only as an special case handle the first
argument as a YAML filename along with a deprecated warning.