Initial implemention of a display for pipelined activity
The TensorCanvas.addActitity() method now takes a optional "wait"
keyword argument. That argument is a dictionary of tracked tensor
numbers and a "delay". For all the specified tensors, the method will
look at all the "points" being tracked and assign the activity to the
latest cycle any "point" was last updated plus the specified "delay".