Allow TensorCanvas.addActivity() "wait=" keyword to be symbolic
The TensorCanvas.addActivity() keyword argument "wait", which is used
for illustrating pipeline delays, can now be symbolic. The keyword
argument is now a dictionary of tensor names (from the tensors given
as __init__() arguments) mapped to delays. For backwards compatibility,
the map key may be the relative offset number of the tensor in the
__init__() call.

Updated the two IPython notebooks that used the "wait" keyword
argument to use symbolic names.