Convert notebooks to universal bootstrapping code
Convert all the notebooks to create their startup environment using a
code cell with a bootstrapping code sequence that will work in a local
environment, the fibertree-docker container or Google Colab.  This is
accomplished using the `fibertree-bootstrap` package, including
downloading it from Github if it does not exist.

The download of the fibertree-bootstrap code itself is accomplished
using the command `python3 -m pip`, since it will work in many
environments.

Toluwa found that using !{sys.executable} is probably even more
general, but that seemed a bit inscrutable.

The environment is established with a call to a method of the
`fibertree-bootstrap` package. That call will establish an environment
for running the fibertree code by setting some global variables and
creating convenience functions. Arguments to the method allow
controlling the environment, e.g., setting the default style for
displaying fibertrees. Finally, that method will also download and
install the fibertree package if it does not exist.

Other changes in this commit include changing a number of uses of
`addFrame()` to `addActivity()`, but many more instances still need to
be changed.