Add/update information/naviation files
Updated the information (README) and naviation (index notebooks). This
includes adding per directory listings and adding links to both a
local notebook and a Google Colab instantiation of the notebook.

The information in the README.md file and the 00-01.index.ipynb files
is identical and exist just to make it convenient to access the
information from different environments, e.g., from github.com or in
Jupyter lab.